+++ client/src/containers/user/userDashboard.js
@@ -1,4 +1,5 @@
 import Items from '../sharedScreens/items'
+import CustomTable from '../../components/customTable'
 function UserDashboard() {
   return (
     <>
@@ -6,6 +7,8 @@ function UserDashboard() {
          Welcome to home
     </div> 
     <Items/>
+    <CustomTable/>
+
     </>
   )
 }
