+++ server/package.json
@@ -18,6 +18,7 @@
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.8.2",
     "multer": "^1.4.5-lts.1",
+    "nodemon": "^2.0.20",
     "react-router-dom": "^6.6.1"
   }
 }
